- moded StopWatch to own file
