+++ storage/Utils/Makefile.am
@@ -9,6 +9,7 @@ libutils_la_SOURCES =					\
 	LoggerImpl.h		LoggerImpl.cc		\
 	AppUtil.cc		AppUtil.h		\
 	Text.cc			Text.h			\
+	StopWatch.cc		StopWatch.h		\
 	Math.cc			Math.h			\
 	FileUtils.cc		FileUtils.h		\
 	Exception.h		Exception.cc		\
