+++ storage/Utils/SystemCmd.cc
@@ -33,7 +33,7 @@
 #include <boost/algorithm/string.hpp>
 
 #include "storage/Utils/ExceptionImpl.h"
-#include "storage/Utils/AppUtil.h"
+#include "storage/Utils/StopWatch.h"
 #include "storage/Utils/LoggerImpl.h"
 #include "storage/Utils/SystemCmd.h"
 #include "storage/Utils/Mockup.h"
